last commit

+++ src/main/java/com/tech/exam/employees/service/EmployeeServiceImpl.java
@@ -31,7 +31,13 @@ public class EmployeeServiceImpl implements IEmployeeService {
   @Transactional
   public void create(EmployeeRequest employee) throws ServiceException {
     try {
+      if (employeeNameExists(employee.getFullName())) {
+        throw new BusinessException(
+            "Employee with name '" + employee.getFullName() + "' already exists");
+      }
       iEmployeeDao.save(employeeMapper.employeeRequestToEmployeeEntity(employee));
+    } catch (BusinessException e) {
+      throw e;
     } catch (Exception e) {
       log.error("Error EmployeeServiceImpl::create");
       throw new ServiceException("Error EmployeeServiceImpl::create");
@@ -45,8 +51,8 @@ public class EmployeeServiceImpl implements IEmployeeService {
       Employee employee = iEmployeeDao.findById(id)
           .orElseThrow(() -> new BusinessException("Employee not found"));
       return employeeMapper.employeeEntityToEmployeeResponse(employee);
+
     } catch (BusinessException e) {
-      log.error("Employee not found", e);
       throw e;
     } catch (Exception e) {
       log.error("Error EmployeeServiceImpl::findAll", e);
@@ -74,7 +80,24 @@ public class EmployeeServiceImpl implements IEmployeeService {
   @Transactional
   public void update(Employee employee) throws ServiceException {
     try {
-      iEmployeeDao.save(employee);
+
+      Employee employeeToUpdate = iEmployeeDao.findById(employee.getId())
+          .orElseThrow(() -> new BusinessException("Employee does not exist"));
+
+      boolean employeeChangeFullName = !employee.getFullName()
+          .equals(employeeToUpdate.getFullName());
+
+      if (employeeChangeFullName && employeeNameExists(employee.getFullName())) {
+        throw new BusinessException(
+            "Employee with name '" + employee.getFullName() + "' already exists");
+      }
+
+      employeeToUpdate.setFullName(employee.getFullName());
+      employeeToUpdate.setFunction(employee.getFunction());
+      iEmployeeDao.save(employeeToUpdate);
+
+    } catch (BusinessException e) {
+      throw e;
     } catch (Exception e) {
       log.error("Error EmployeeServiceImpl::update", e);
       throw new ServiceException("Error EmployeeServiceImpl::update", e);
@@ -85,7 +108,13 @@ public class EmployeeServiceImpl implements IEmployeeService {
   @Transactional
   public void delete(Long id) throws ServiceException {
     try {
+      if (!iEmployeeDao.findById(id).isPresent()) {
+        throw new BusinessException("Employee does not exist");
+      }
       iEmployeeDao.deleteById(id);
+
+    } catch (BusinessException e) {
+      throw e;
     } catch (Exception e) {
       log.error("Error EmployeeServiceImpl::delete", e);
       throw new ServiceException("Error EmployeeServiceImpl::delete", e);
@@ -96,11 +125,19 @@ public class EmployeeServiceImpl implements IEmployeeService {
   @Transactional
   public void setBoss(Long idBoss, Long idEmployee) throws ServiceException {
     try {
+      if (idBoss.equals(idEmployee)) {
+        throw new BusinessException("You can't be your own boss.");
+      }
+
       Employee employee = iEmployeeDao.findById(idEmployee)
           .orElseThrow(() -> new BusinessException("Employee not found"));
 
       Employee boss = iEmployeeDao.findById(idBoss)
           .orElseThrow(() -> new BusinessException("Boss not found"));
+
+      if(boss.getBoss() != null && boss.getBoss().getId().equals(idEmployee)){
+        throw new BusinessException("You can't be boss of your boss.");
+      }
 
       employee.setBoss(boss);
       iEmployeeDao.save(employee);
@@ -110,6 +147,11 @@ public class EmployeeServiceImpl implements IEmployeeService {
       log.error("Error EmployeeServiceImpl::setBoss", e);
       throw new ServiceException("Error EmployeeServiceImpl::setBoss", e);
     }
+  }
+
+  private boolean employeeNameExists(String fullName){
+    return iEmployeeDao.findByFullName(fullName)
+        .isPresent();
   }
 
 }
